refactor: encapsulate mutations in TableActions component
