+++ packages/apps/spaces/app/organizations/src/components/Actions/Actions.tsx
@@ -7,31 +7,65 @@ import { Organization } from '@graphql/types';
 import { Copy07 } from '@ui/media/icons/Copy07';
 import { Archive } from '@ui/media/icons/Archive';
 import { ButtonGroup } from '@ui/form/ButtonGroup';
-import { TableInstance, RowSelectionState } from '@ui/presentation/Table';
+import { TableInstance } from '@ui/presentation/Table';
 import { ConfirmDeleteDialog } from '@ui/overlay/AlertDialog/ConfirmDeleteDialog';
 
+import { useOrganizationsPageMethods } from '../../hooks/useOrganizationsPageMethods';
+
 interface TableActionsProps {
-  isArchiving: boolean;
+  allOrganizationsIds: string[];
   table: TableInstance<Organization>;
-  onMergeOrganizations: (
-    targetIndex: string,
-    selection: RowSelectionState,
-  ) => void;
-  onArchiveOrganizations: (selection: RowSelectionState) => void;
 }
 
 export const TableActions = memo(
-  ({
-    table,
-    isArchiving,
-    onMergeOrganizations,
-    onArchiveOrganizations,
-  }: TableActionsProps) => {
+  ({ table, allOrganizationsIds }: TableActionsProps) => {
     const { isOpen, onOpen, onClose } = useDisclosure();
     const [targetIndex, setTargetIndex] = useState<string | null>(null);
+    const { hideOrganizations, mergeOrganizations } =
+      useOrganizationsPageMethods();
 
     const selection = table.getState().rowSelection;
     const selectCount = Object.keys(selection).length;
+
+    const handleMergeOrganizations = () => {
+      const primaryId = (allOrganizationsIds as string[])[Number(targetIndex)];
+      const selectedIds = Object.keys(selection).map(
+        (k) => (allOrganizationsIds as string[])[Number(k)],
+      );
+      const mergeIds = selectedIds.filter((id) => id !== primaryId);
+
+      if (!primaryId || !mergeIds.length) return;
+
+      mergeOrganizations.mutate(
+        {
+          primaryOrganizationId: primaryId,
+          mergedOrganizationIds: mergeIds,
+        },
+        {
+          onSuccess: () => {
+            table.resetRowSelection();
+          },
+        },
+      );
+    };
+
+    const handleHideOrganizations = () => {
+      const selectedIds = Object.keys(selection)
+        .map((k) => (allOrganizationsIds as string[])[Number(k)])
+        .filter(Boolean);
+
+      hideOrganizations.mutate(
+        {
+          ids: selectedIds,
+        },
+        {
+          onSuccess: () => {
+            onClose();
+            table.resetRowSelection();
+          },
+        },
+      );
+    };
 
     useEffect(() => {
       if (selectCount === 1) {
@@ -70,10 +104,7 @@ export const TableActions = memo(
                 _hover={{
                   bg: 'gray.800',
                 }}
-                onClick={() => {
-                  onMergeOrganizations(targetIndex as string, selection);
-                  table.resetRowSelection();
-                }}
+                onClick={handleMergeOrganizations}
               >
                 {`Merge these ${selectCount}`}
               </Button>
@@ -85,13 +116,9 @@ export const TableActions = memo(
           isOpen={isOpen}
           icon={<Archive />}
           onClose={onClose}
-          isLoading={isArchiving}
           confirmButtonLabel={'Archive'}
-          onConfirm={() => {
-            onArchiveOrganizations(selection);
-            onClose();
-            table.resetRowSelection();
-          }}
+          onConfirm={handleHideOrganizations}
+          isLoading={hideOrganizations.isLoading}
           label={`Archive selected ${
             selectCount === 1 ? 'organization' : 'organizations'
           }?`}
@@ -102,6 +129,9 @@ export const TableActions = memo(
   (prev, next) => {
     const prevCount = Object.keys(prev.table.getState().rowSelection).length;
     const nextCount = Object.keys(next.table.getState().rowSelection).length;
-    return prevCount !== nextCount;
+    const prevIdsCount = prev.allOrganizationsIds.length;
+    const nextIdsCount = next.allOrganizationsIds.length;
+
+    return prevCount !== nextCount || prevIdsCount !== nextIdsCount;
   },
 );
